Day17 Whisper Desktop/requirements.txt
PySide6>=6.6
numpy>=1.23
sounddevice>=0.4
soundfile>=0.12
librosa>=0.10
faster-whisper>=1.0
huggingface_hub>=0.23
imageio-ffmpeg>=0.4
srt>=3.5
webvtt-py>=0.4
noisereduce>=3.0
concurrent-log-handler>=0.9
openai>=1.40
# optional via 設定頁「安裝選配」或自動安裝：
# torch>=2.1
# pyannote.audio>=3.1
# demucs>=4.0
# 語言修正：
# opencc
